WIP: Connect EditorActions
