+++ ui/src/components/activity/ActivityHeader/ActivityHeader.module.css
@@ -17,14 +17,3 @@
   flex-direction: row;
   align-items: center;
 }
-
-.lastSave {
-  margin-right: 1em;
-  font-size: 0.8em;
-  color: rgba(0, 0, 0, 0.4);
-}
-
-.error {
-  color: #db2828;
-  font-weight: bold;
-}
