+++ ui/src/components/editor/EditorActions/ImportBarcodesButton/index.js
