+++ ui/src/components/editor/EditorActions/index.js
@@ -1,0 +1,2 @@
+import { EditorActions } from './connect';
+export { EditorActions };
