+++ ui/src/store/designExperiment.js
@@ -73,13 +73,6 @@ const designExperiment = createSlice({
     setStepThreeComplete(state, action) {
       state.steps.stepThreeCompleted = true;
     },
-    setSaveStatus(state, action) {
-      const { saveStatus } = action.payload;
-      state.saveStatus = saveStatus;
-      if (saveStatus === 'SUCCESS') {
-        state.lastSaveTime = Date.now();
-      }
-    },
   },
 });
